refactor: remove `registration.cpp` hack
Use automatic registration to save some effort.
In a separate commit, will convert peppx back to pepp.
