+++ .github/workflows/ci.yml
@@ -178,7 +178,7 @@ jobs:
       MAC_DEVELOPER_NAME: ${{ secrets.MAC_DEVELOPER_NAME }}
   deploy_pages:
     name: Trigger deployment to github pages
-    if: startsWith(needs.branch_info.outputs.not_default, 'false')
+    # if: startsWith(needs.branch_info.outputs.not_default, 'false')
     permissions:
       contents: read
       pages: write
