+++ bin/gui/main.qml
@@ -21,16 +21,16 @@ import QtQuick.Controls
 import QtQuick.Layouts
 import Qt.labs.qmlmodels
 import QtCore
-import "qrc:/ui/about" as About
-import "qrc:/ui/components" as Comp
-import "qrc:/ui/memory/hexdump" as Memory
-import "qrc:/ui/cpu" as Cpu
-import "qrc:/ui/text/editor" as Editor
-import "qrc:/ui/project" as Project
-import "qrc:/ui/preferences" as Pref
-import "qrc:/ui/builtins" as Builtins
-import edu.pepp 1.0
-import "qrc:/ui/menu" as Menu
+import "qrc:/edu/peppx/about" as About
+import "qrc:/edu/peppx/components" as Comp
+import "qrc:/edu/peppx/memory/hexdump" as Memory
+import "qrc:/edu/peppx/cpu" as Cpu
+import "qrc:/edu/peppx/text/editor" as Editor
+import "qrc:/edu/peppx/project" as Project
+import "qrc:/edu/peppx/preferences" as Pref
+import "qrc:/edu/peppx/builtins" as Builtins
+import edu.peppx 1.0
+import "qrc:/edu/peppx/menu" as Menu
 //  Native menu for apple, linux, and windows
 import Qt.labs.platform as Labs
 
@@ -74,7 +74,6 @@ ApplicationWindow {
             button: Theme.base.background
         }
     }
-
     property var currentProject: null
     // Used to expose actions to inner area.
     property var actionRef: actions
@@ -127,6 +126,7 @@ ApplicationWindow {
     }
 
     Component.onCompleted: {
+        console.log(ApplicationPreferences)
         // Allow welcome mode to create a new project, and switch to it on creation.
         welcome.addProject.connect(pm.onAddProject)
         welcome.addProject.connect(() => switchToProject(pm.count - 1))
@@ -241,11 +241,11 @@ ApplicationWindow {
             }
             if (PlatformDetector.isWASM) {
                 props["window"] = window
-                setSource("qrc:/ui/menu/QMLMainMenu.qml", props)
+                setSource("qrc:/edu/peppx/menu/QMLMainMenu.qml", props)
             } else
                 // Auto-recurses on "parent" to find "window" of correct type.
                 // If explicitly set, the menu bar will not render until hovered over.
-                setSource("qrc:/ui/menu/NativeMainMenu.qml", props)
+                setSource("qrc:/edu/peppx/menu/NativeMainMenu.qml", props)
         }
         onLoaded: {
             if (PlatformDetector.isWASM)
