+++ lib/builtins/Figure2.qml
@@ -18,8 +18,8 @@
 import QtQuick
 import QtQuick.Controls
 import QtQuick.Layouts
-import edu.pepp 1.0
-import "qrc:/ui/text/editor" as Editor
+import edu.peppx 1.0
+import "qrc:/edu/peppx/text/editor" as Editor
 
 Item {
     id: wrapper
