+++ lib/components/charcheck.hpp
@@ -2,9 +2,12 @@
 
 #include <QFont>
 #include <QObject>
+#include <QtQmlIntegration>
 
 class CharCheck : public QObject {
   Q_OBJECT
+  QML_ELEMENT
+
 public:
   explicit CharCheck(QObject *parent = nullptr);
   Q_INVOKABLE bool isCharSupported(const QString &character, const QFont &font);
