+++ lib/menu/QMLMainMenu.qml
@@ -3,7 +3,7 @@ import QtQuick.Controls
 import QtQuick.Layouts
 import "./"
 // For PlatformDetector
-import edu.pepp 1.0
+import edu.peppx 1.0
 
 MenuBar {
     id: wrapper
