+++ lib/project/aproject.hpp
@@ -1,14 +1,10 @@
 #pragma once
 #include <QQmlEngine>
 #include <QStringListModel>
-#include <deque>
+#include <QtQmlIntegration>
 #include <qabstractitemmodel.h>
 #include <targets/pep10/isa3/system.hpp>
 #include "builtins/constants.hpp"
-#include "cpu/registermodel.hpp"
-#include "cpu/statusbitmodel.hpp"
-#include "memory/hexdump/rawmemory.hpp"
-#include "utils/opcodemodel.hpp"
 
 namespace sim {
 namespace trace2 {
@@ -28,6 +24,9 @@ enum class Features : int {
 
 class DebugEnableFlags : public QObject {
   Q_OBJECT
+  QML_NAMED_ELEMENT(DebugEnableFlags)
+  QML_UNCREATABLE("Error: Only enums")
+
 public:
   explicit DebugEnableFlags(QObject *parent = nullptr);
   enum Value {
@@ -43,6 +42,9 @@ public:
 
 class StepEnableFlags : public QObject {
   Q_OBJECT
+  QML_NAMED_ELEMENT(StepEnableFlags)
+  QML_UNCREATABLE("Error: Only enums")
+
 public:
   explicit StepEnableFlags(QObject *parent = nullptr);
   enum Value {
@@ -80,37 +82,4 @@ public:
 
 private:
   const project::Environment _env;
-};
-
-class Pep10_ISA;
-class Pep10_ASMB;
-// Factory to ensure class invariants of project are maintained.
-// Must be a singleton to call methods on it.
-// Can't seem to call Q_INVOKABLE on an uncreatable type.
-class ProjectModel : public QAbstractListModel {
-  Q_OBJECT
-  Q_PROPERTY(int count READ _rowCount NOTIFY rowCountChanged)
-public:
-  enum class Roles {
-    ProjectRole = Qt::UserRole + 1,
-  };
-  Q_ENUM(Roles);
-  // Q_INVOKABLE ISAProject *isa(utils::Architecture::Value arch, project::Features features);
-  explicit ProjectModel(QObject *parent = nullptr) : QAbstractListModel(parent){};
-  // Helper to expose rowCount as a property to QML.
-  int _rowCount() const { return rowCount({}); }
-  int rowCount(const QModelIndex &parent) const override;
-  QVariant data(const QModelIndex &index, int role) const override;
-  Q_INVOKABLE Pep10_ISA *pep10ISA(QVariant delegate);
-  Q_INVOKABLE Pep10_ASMB *pep10ASMB(QVariant delegate, builtins::Abstraction abstraction);
-  bool removeRows(int row, int count, const QModelIndex &parent) override;
-  bool moveRows(const QModelIndex &sourceParent, int sourceRow, int count, const QModelIndex &destinationParent,
-                int destinationChild) override;
-  QHash<int, QByteArray> roleNames() const override;
-  Q_INVOKABLE QString describe(int index) const;
-signals:
-  void rowCountChanged(int);
-
-private:
-  std::deque<std::unique_ptr<QObject>> _projects = {};
 };
