+++ lib/text/editor/object.hpp
@@ -1,9 +1,12 @@
 #pragma once
 #include <QObject>
+#include <QtQmlIntegration>
 
 class ObjectUtilities : public QObject {
   Q_OBJECT
   Q_PROPERTY(int bytesPerRow READ bytesPerRow WRITE setBytesPerRow NOTIFY bytesPerRowChanged)
+  QML_NAMED_ELEMENT(ObjectUtilities)
+
 public:
   explicit ObjectUtilities(QObject *parent = nullptr);
   Q_INVOKABLE static bool valid(int key);
