+++ lib/utils/sequenceconverter.hpp
@@ -8,6 +8,9 @@
 namespace utils {
 class SequenceConverter : public QObject {
   Q_OBJECT
+  QML_SINGLETON
+  QML_ELEMENT
+
 public:
   explicit SequenceConverter(QObject *parent = nullptr);
   Q_INVOKABLE QString toNativeText(const QVariant &sequence);
